Round number to nearest integer
Using the Math.round() function to round the random number enables the program to avoid edge
cases where the function returns "undefined" due to the floating point giving the number a
value in between the options specified in the 'if' block.

Another option would have been to change the operators.